Remove separate folder for passthrough
